requirements.txt
backports.ssl-match-hostname==3.4.0.2
BeautifulSoup==3.2.1
boto==2.38.0
certifi==2015.4.28
cssselect==0.9.1
eventlet==0.17.4
feedparser==5.2.1
gevent==1.0.2
goose-extractor==1.0.25
greenlet==0.4.7
gunicorn==19.3.0
ipython==3.1.0
jieba==0.36.2
Pillow==2.8.2
pymongo==3.0.2
requests==2.7.0
tornado==4.0.2
wheel==0.24.0
zope.interface==4.1.2
urllib3==1.14
flask==0.10.1
flask-restful==0.3.5
elasticsearch==2.3.0
termcolor==1.1.0
pyfiglet==0.7.4
